Route validation
- Continue wto work on route validation
- Remove bad use statemens
