+++ app/Http/Controllers/ItemSubtypeView.php
@@ -12,6 +12,7 @@ use App\Request\Route;
 use App\Response\Header\Headers;
 use App\Response\Pagination as UtilityPagination;
 use App\Models\Transformers\ItemSubtype as ItemSubtypeTransformer;
+use App\Response\Responses;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Config;
 
@@ -26,7 +27,9 @@ class ItemSubtypeView extends Controller
 
     public function index($item_type_id): JsonResponse
     {
-        Route\Validate::itemType($item_type_id);
+        if (Route\Validate\ItemType::existsToUserForViewing($item_type_id) === false) {
+            Responses::notFound(trans('entities.item-subtype'));
+        }
 
         $cache_control = new Cache\Control();
         $cache_control->setTtlOneYear();
@@ -86,7 +89,9 @@ class ItemSubtypeView extends Controller
 
     public function show($item_type_id, $item_subtype_id): JsonResponse
     {
-        Route\Validate::itemSubType($item_type_id, $item_subtype_id);
+        if (Route\Validate\ItemSubtype::existsToUserForViewing($item_type_id, $item_subtype_id) === false) {
+            Responses::notFound(trans('entities.item-subtype'));
+        }
 
         $subtype = (new ItemSubtype())->single(
             (int) $item_type_id,
@@ -94,7 +99,7 @@ class ItemSubtypeView extends Controller
         );
 
         if ($subtype === null) {
-            return \App\Response\Responses::notFound(trans('entities.item-subtype'));
+            return Responses::notFound(trans('entities.item-subtype'));
         }
 
         $headers = new Header();
@@ -109,7 +114,9 @@ class ItemSubtypeView extends Controller
 
     public function optionsIndex($item_type_id): JsonResponse
     {
-        Route\Validate::itemType($item_type_id);
+        if (Route\Validate\ItemType::existsToUserForViewing($item_type_id) === false) {
+            Responses::notFound(trans('entities.item-subtype'));
+        }
 
         $response = new ItemSubtypeCollection(['view'=> $this->user_id !== null]);
 
@@ -118,10 +125,9 @@ class ItemSubtypeView extends Controller
 
     public function optionsShow($item_type_id, $item_subtype_id): JsonResponse
     {
-        Route\Validate::itemSubType(
-            $item_type_id,
-            $item_subtype_id
-        );
+        if (Route\Validate\ItemSubtype::existsToUserForViewing($item_type_id, $item_subtype_id) === false) {
+            Responses::notFound(trans('entities.item-subtype'));
+        }
 
         $response = new ItemSubtypeItem(['view'=> $this->user_id !== null]);
