+++ app/Listeners/CaptureAndSendRequestError.php
@@ -4,8 +4,6 @@ namespace App\Listeners;
 
 use App\Events\RequestError;
 use App\Mail\RequestError as RequestErrorMail;
-use Illuminate\Queue\InteractsWithQueue;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Mail;
