+++ app/Models/PermittedUser.php
@@ -26,7 +26,7 @@ class PermittedUser extends Model
      * @param integer $resource_type_id
      * @param integer $user_id
      *
-     * @return ResourceTypeAccess|null
+     * @return ResourceAccess|null
      */
     public function instance(int $resource_type_id, int $user_id): ?PermittedUser
     {
