+++ app/Models/ResourceTypeItem/Summary/AllocatedExpense.php
@@ -6,7 +6,6 @@ namespace App\Models\ResourceTypeItem\Summary;
 use App\Interfaces\ResourceTypeItem\ISummaryModelCategories;
 use App\Interfaces\ResourceTypeItem\ISummaryModel;
 use App\Models\Clause;
-use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use Illuminate\Support\Facades\DB;
