+++ app/Request/Route/Validate/ItemSubtype.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Request\Route\Validator;
+namespace App\Request\Route\Validate;
 
-use App\Models\ResourceTypeAccess;
+use App\Models\ResourceAccess;
 
 /**
  * Validate the route params to an item subtype
@@ -23,7 +23,7 @@ class ItemSubtype
     public static function existsToUserForViewing($item_type_id, $item_subtype_id): bool
     {
         return !(
-            (new ResourceTypeAccess())->itemSubTypeExistsToUser((int) $item_type_id, (int) $item_subtype_id) === false
+            (new ResourceAccess())->itemSubTypeExistsToUser((int) $item_type_id, (int) $item_subtype_id) === false
         );
     }
 }
