+++ app/Request/Route/Validate/Queue.php
@@ -1,30 +1,28 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Request\Route\Validator;
+namespace App\Request\Route\Validate;
 
-use App\Models\ResourceTypeAccess;
+use App\Models\ResourceAccess;
 
 /**
- * Validate the route params to a currency
- *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class Currency
+class Queue
 {
     /**
      * Validate that the user is able to view the requested currency
      *
-     * @param string|int $currency_id
+     * @param string|int $queue_id
      *
      * @return boolean
      */
-    public static function existsToUserForViewing($currency_id): bool
+    public static function existsToUserForViewing($queue_id): bool
     {
         return !(
-            (new ResourceTypeAccess())->currencyExistsToUser((int) $currency_id) === false
+            (new ResourceAccess())->queueExistsToUser((int) $queue_id) === false
         );
     }
 }
